move code/preview to @app/design-to-code
