+++ app/__plugin__init__/index.ts
@@ -1,5 +1,5 @@
 // DO NOT REMOVE THIS LINE
-import "../lib/pages/code/__plugin";
+import "@app/design-to-code/__plugin";
 import "@app/data-mapper/__plugin";
 import "@app/design-lint/__plugin";
 import "@app/design-text-code-syntax-highlight/__plugin";
