+++ packages/app-design-to-code/package.json
@@ -1,0 +1,5 @@
+{
+    "name": "@app/design-to-code",
+    "version": "0.0.0",
+    "private": false
+}
